first iteration for antcolony works

+++ src/Matrix.java
@@ -5,17 +5,20 @@ import java.util.Scanner;
 
 public class Matrix {
 	private float[][] matrix_ = new float[16][30];
-	
+
 	public void loadMatrixFromFile(String path) throws FileNotFoundException{
 		Scanner scanner = new Scanner(new File(path));
 		for (int row = 0; row < 30; row++){
 			for (int column = 0; column < 16; column++){
-				matrix_[column][row] = scanner.nextFloat();
+				matrix_[column][row] = scanner.nextInt();
 			}
 		}
 	}
-	
+
 	public float getValue(int column, int row){
+
 		return matrix_[column][row];
+
+
 	}
 }
